Fix Error and Warning from the compilation after merging branch
Fix compilation errors from the merging of the branches to the master:
several namespace were missing.
Fix also main compilation warning related to signed int and unsigned
int comparisons, unused variable and more importantly several variable
shadowing.

+++ src/FullRecoEvent.cc
@@ -66,15 +66,15 @@ void FullRecoEvent::Clear(int toclean)
   FragmentPID = -999;
 
   for(auto x : FiberTrackCont)
-    for(int i=0; i<x.second.size(); ++i)
+    for(size_t i=0; i<x.second.size(); ++i)
       delete x.second[i];
 
-  for(int i = 0; i < FiberHitClCont.size(); ++i)
+  for(size_t i = 0; i < FiberHitClCont.size(); ++i)
     {
-      for(int j = 0; j < FiberHitClCont[i].size(); ++j)
+      for(size_t j = 0; j < FiberHitClCont[i].size(); ++j)
         {
-          int num = (int)FiberHitClCont[i][j].size();
-          for(int k = 0; k < num; ++k)
+          size_t num = FiberHitClCont[i][j].size();
+          for(size_t k = 0; k < num; ++k)
             delete FiberHitClCont[i][j][k];
         }
     }
